updated user model
